- Solved error when return string on format double

+++ Weather/Weather.cs
@@ -19,7 +19,7 @@ namespace Weather
 	public class Main
 	{
 		public double temp { get; set; }
-		public int pressure { get; set; }
+		public double pressure { get; set; }
 		public int humidity { get; set; }
 		public double temp_min { get; set; }
 		public double temp_max { get; set; }
@@ -28,7 +28,7 @@ namespace Weather
 	public class Wind
 	{
 		public double speed { get; set; }
-		public int deg { get; set; }
+		public double deg { get; set; }
 	}
 
 	public class Clouds
